+++ CHANGELOG.md
@@ -43,6 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Consensus Breaking Changes
 
+* (x/intent) [#139](https://github.com/warden-protocol/wardenprotocol/pull/139) Ability for modules to dynamically resolve variables on Action creation
+    * x/warden can now resolve `warden.space.owners` in Intent definitions into the list of owners of the space
+
 ### Features
 
 ### Bug Fixes